Change CertificateAuthorityActivation resource Secret annotations to spec field (#64)
Description of changes:
Added Spec field `completeCertificateChainOutput` and removed use of annotations for CertificateAuthorityActivation resource. The complete certificate chain will now be written to the Secret specified by the `completeCertificateChainOutput` field.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.